Move receive::common::exclusive to receive::common::v1
