+++ payjoin/src/core/receive/common/v1/error.rs
